move multipliers to their own file
