+++ x/incentive/types/params.go
@@ -7,20 +7,12 @@ import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/params"
 
 	tmtime "github.com/tendermint/tendermint/types/time"
 
 	cdptypes "github.com/kava-labs/kava/x/cdp/types"
 	kavadistTypes "github.com/kava-labs/kava/x/kavadist/types"
-)
-
-// Valid reward multipliers
-const (
-	Small  MultiplierName = "small"
-	Medium MultiplierName = "medium"
-	Large  MultiplierName = "large"
 )
 
 // Parameter keys and default values
@@ -341,78 +333,4 @@ func (mrps MultiRewardPeriods) Validate() error {
 	}
 
 	return nil
-}
-
-// Multiplier amount the claim rewards get increased by, along with how long the claim rewards are locked
-type Multiplier struct {
-	Name         MultiplierName `json:"name" yaml:"name"`
-	MonthsLockup int64          `json:"months_lockup" yaml:"months_lockup"`
-	Factor       sdk.Dec        `json:"factor" yaml:"factor"`
-}
-
-// NewMultiplier returns a new Multiplier
-func NewMultiplier(name MultiplierName, lockup int64, factor sdk.Dec) Multiplier {
-	return Multiplier{
-		Name:         name,
-		MonthsLockup: lockup,
-		Factor:       factor,
-	}
-}
-
-// Validate multiplier param
-func (m Multiplier) Validate() error {
-	if err := m.Name.IsValid(); err != nil {
-		return err
-	}
-	if m.MonthsLockup < 0 {
-		return fmt.Errorf("expected non-negative lockup, got %d", m.MonthsLockup)
-	}
-	if m.Factor.IsNegative() {
-		return fmt.Errorf("expected non-negative factor, got %s", m.Factor.String())
-	}
-
-	return nil
-}
-
-// String implements fmt.Stringer
-func (m Multiplier) String() string {
-	return fmt.Sprintf(`Claim Multiplier:
-	Name: %s
-	Months Lockup %d
-	Factor %s
-	`, m.Name, m.MonthsLockup, m.Factor)
-}
-
-// Multipliers slice of Multiplier
-type Multipliers []Multiplier
-
-// Validate validates each multiplier
-func (ms Multipliers) Validate() error {
-	for _, m := range ms {
-		if err := m.Validate(); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// String implements fmt.Stringer
-func (ms Multipliers) String() string {
-	out := "Claim Multipliers\n"
-	for _, s := range ms {
-		out += fmt.Sprintf("%s\n", s)
-	}
-	return out
-}
-
-// MultiplierName name for valid multiplier
-type MultiplierName string
-
-// IsValid checks if the input is one of the expected strings
-func (mn MultiplierName) IsValid() error {
-	switch mn {
-	case Small, Medium, Large:
-		return nil
-	}
-	return sdkerrors.Wrapf(ErrInvalidMultiplier, "invalid multiplier name: %s", mn)
 }
